move rate limiter to it's own class
